requirements.txt
aiogram~=3.4.1
fluent.runtime==0.4.0
SQLAlchemy~=2.0.27
aiosqlite~=0.20.0
alembic~=1.13.1
Mako~=1.3.2
greenlet~=3.0.3
markupsafe~=2.1.5
cachetools~=5.3.3
pydantic~=2.5.3
PyYAML~=6.0.1